Demand an editor configured:
Without any check for an editor, the processing part tries to load
an editor, fails, has a warning, and assumes everything was fine
and the file just wasn't edited.

This verifies that there's an editor or tells you what to set in
order to make sure you have one.

This shouldn't actually be here but should be at a proper place
that verifies configuration options. Right now it says that the
compilation fails, because it's in a `BEGIN` block.